Restructures project structure and adjusts package's __init__ files and import statements.
