+++ file_load/file.py
@@ -3,9 +3,9 @@ import time
 from pathlib import Path
 import logging
 from file_load.file_error import FileError
-from utils.log_utils import setup_logger_common
+from utils import setup_logger_common
 from utils import global_const as gc
-from .file_utils import StudyConfig
+from file_load import StudyConfig
 from csv import reader
