+++ file_load/rawdata_text.py
@@ -1,6 +1,5 @@
-from utils import global_const as gc
 from file_load import File
-from utils import ConfigData
+
 
 class RawData_Text(File):
