+++ submission.py
@@ -1,11 +1,10 @@
-from file_load import MetaFileText, MetaFileExcel
 from pathlib import Path
 import sys
 import os
 from os import walk
 import time
 import traceback
-from utils.log_utils import setup_logger_common, deactivate_logger_common
+from utils import setup_logger_common, deactivate_logger_common
 from utils import ConfigData
 from utils import global_const as gc
 from utils import send_email as email
