+++ utils/__init__.py
@@ -1,5 +1,5 @@
-from .db_access import *
 from .global_const import *
 from .log_utils import *
 from .configuration import *
+from .db_access import *
 from .send_email import *
